Rename "student-overview-new/" -> "student-list/"
